# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: vcare Identity Service API
  version: 1.0.0
  summary: Accounts, authentication, sessions, account status, and service-to-service tokens for vcare.
  description: |
    Identity owns **who someone is and whether they may act**.

    - Public listener (`PORT`, default `3000`): `/api/*`, `/.well-known/jwks.json`.
    - Internal listener (`INTERNAL_PORT`, default `3100`): `/internal/*`. Bound to the private interface only;
      ingress never routes `/internal`. Every `/internal/*` operation except `/internal/auth/token` and
      `/internal/health/*` requires a **service token** (`typ=service`); a user token (even an admin's) gets
      `401 ServiceTokenRequired`.
    - User access tokens are EdDSA (Ed25519) JWTs, 15 minutes, sent only as `Authorization: Bearer`.
      Consumers verify them locally against `GET /.well-known/jwks.json`.
    - Refresh tokens are opaque, 30 days, delivered only via the `vcare_rt` httpOnly cookie
      (`Path=/api/auth`), rotated on every use with family-wide reuse detection.
    - Every response carries `X-Request-Id`. Every error uses the single `ErrorEnvelope`.
    - Every list is cursor-paginated (`cursor`, `limit` 1..100, default 20).

    Vendor extensions on every operation: `x-roles` (roles allowed by the RBAC policy; `public` = no
    principal; `service` = service token), `x-ownership` (`none` | `self` | `refresh-family`),
...
  contact:
    name: identity-team
  license:
    name: Proprietary
...
servers:
...
security:
  - bearerUser: []
...
paths:
...
  /api/auth/register/start:
    post:
      operationId: startRegistration
      tags: [auth]
      summary: Start registration — email a 6-digit code (ADR 0006)
      description: |
        Always returns `202` with no body, whether or not the email already belongs to a live account (no
        account enumeration). Unknown email: invalidates earlier open challenges for that email, creates a
        `registration_challenges` row, and enqueues an outbox job that emails a 6-digit code (generated by the
        worker at send time, valid 10 minutes from send, max 5 attempts; ADR 0007). Known live email: no
        challenge is created; an "account exists" notice is emailed instead. Email delivery never fails the
        request. A live account is an account that is not deleted, whatever its status. Calling this again
        is the resend; replaying the same `Idempotency-Key` does not send another email — a resend uses a new
        key or none. Rate limits: 3/h per email, 5/h per IP.
      security: []
      x-roles: [public]
      x-ownership: none
      x-error-codes: [ValidationFailed, Conflict, IdempotencyConflict, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterStartRequest'
      responses:
        '202':
          description: Accepted. Identical for known and unknown emails.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/register/complete:
    post:
      operationId: completeRegistration
      tags: [auth]
      summary: Complete registration with the emailed code and create the account (ADR 0006)
      description: |
        In one transaction: loads the latest open challenge for the email, compares
        `HMAC-SHA256(OTP_PEPPER, code)` in constant time, increments `attempts`, and on success sets
        `consumed_at` and creates the user with `email_verified_at = now()`. `role` is patient or doctor only
        (admins are never created here); patients start `status=active`, doctors `status=pending`.
        A wrong, expired, or consumed code, a missing challenge, or the 5th failed attempt (which invalidates
        the challenge) returns `400 ValidationFailed` with `details[].field = "code"`. `409 Conflict` only when
        a concurrent registration for the same email won the race after ownership was proven. Does not log the
        user in. `Idempotency-Key` is **required**; same key + same body replays the original response, same
        key + different body returns `422 IdempotencyConflict`. Rate limit: 10/h per IP.
      security: []
      x-roles: [public]
      x-ownership: none
      x-error-codes: [ValidationFailed, Conflict, IdempotencyConflict, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCompleteRequest'
      responses:
        '201':
          description: Account created with a verified email.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      description: |
        Returns a 15-minute access token in the body and sets the rotating refresh token in the `vcare_rt`
        cookie (new refresh-token family). Unknown email and wrong password return the same
        `401 InvalidCredentials` with equalized timing. `suspended` accounts cannot log in
        (`403 AccountSuspended`). `pending` and `rejected` accounts can (ADR 0004); their token carries
        `status=pending` / `status=rejected`. Legacy bcrypt hashes are rehashed to argon2id on success.
        `Idempotency-Key` is not accepted on this route: a retried login simply creates another session.
        Rate limits: 5/min per IP+email, 20/min per IP.
      security: []
      x-roles: [public]
      x-ownership: none
      x-error-codes: [ValidationFailed, InvalidCredentials, AccountSuspended, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated. Refresh token delivered only via Set-Cookie.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
            Set-Cookie: { $ref: '#/components/headers/SetRefreshCookie' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LoginResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '403': { $ref: '#/components/responses/AccountStateForbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/refresh:
    post:
      operationId: refreshAccessToken
      tags: [auth]
      summary: Rotate the refresh token and issue a new access token
      description: |
        Reads the `vcare_rt` cookie. In one transaction: revokes the presented token
        (`revoked_reason='rotated'`, `replaced_by_id`) and issues a new token in the same family.
        Presenting an already-rotated token revokes the **entire family** (`reuse_detected`) and returns
        `401 RefreshTokenReused`.

        **Grace window (ADR 0005):** if the presented token was rotated less than `REFRESH_REUSE_GRACE_SECONDS`
        (10 s) ago and its successor is still live (not rotated or revoked), the response is
        `401 RefreshTokenInvalid`, the family is **not** revoked, and **no** clearing `Set-Cookie` is sent (the
        shared cookie jar already holds the successor). Clients retry once, then call logout if it fails again.

        The user is re-read: `suspended` revokes the family and returns `403 AccountSuspended`. `pending` and
        `rejected` accounts may refresh (ADR 0004); the new token carries their status. Every other 401/403
        clears the cookie. Rate limit: 30/min per family. Budget p95 < 50 ms.
      security:
        - refreshCookie: []
      x-roles: [public]
      x-ownership: refresh-family
      x-error-codes: [RefreshTokenInvalid, RefreshTokenReused, AccountSuspended, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: New access token; rotated refresh token in Set-Cookie.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
            Set-Cookie: { $ref: '#/components/headers/SetRefreshCookie' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AccessTokenResponse' }
        '401': { $ref: '#/components/responses/RefreshUnauthorized' }
        '403': { $ref: '#/components/responses/RefreshAccountSuspended' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out — revoke the refresh-token family and clear the cookie
      description: |
        Revokes the family of the presented `vcare_rt` token (`revoked_reason='logout'`) and clears the cookie.
        Always returns 204, including when the cookie is absent, unknown, or already revoked, so clients can
        call it unconditionally. The already-issued access token remains valid until it expires (at most 15 min).
        No principal is required; when a cookie is present only its own family is revoked.
      security:
        - refreshCookie: []
        - {}
      x-roles: [public]
      x-ownership: refresh-family
      x-error-codes: [InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '204':
          description: Logged out; cookie cleared.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
            Set-Cookie: { $ref: '#/components/headers/ClearRefreshCookie' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password-reset email
      description: |
        Always returns 204 regardless of whether the email exists. When a live account matches (an account
        that is not deleted, whatever its status), earlier unused codes are invalidated and a 30-minute
        single-use 6-digit code is emailed asynchronously.
        Rate limit: 3/h per email.
      security: []
      x-roles: [public]
      x-ownership: none
      x-error-codes: [ValidationFailed, Conflict, IdempotencyConflict, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '204':
          description: Accepted.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password using a reset code
      description: |
        Consumes a 6-digit password-reset code (valid 30 minutes from send, single use, **max 5 attempts** —
        the 5th failure invalidates it). `email` and `code` must belong to the same account. On success the
        new argon2id hash is set, the code is marked `used_at`, and **all** refresh-token families for the
        user are revoked in the same transaction. An unknown email, a wrong, never-sent, expired, used,
        superseded, or attempt-exhausted code all return the identical `400 ValidationFailed` with
        `details[].field = "code"`, so the route cannot be used to enumerate accounts. `suspended` accounts
        may reset their password but still cannot log in; `rejected` and `pending` accounts can log in
        (ADR 0004). The account's status is unchanged.
        Rate limits: 5/h per email, 10/h per IP.
      security: []
      x-roles: [public]
      x-ownership: none
      x-error-codes: [ValidationFailed, Conflict, IdempotencyConflict, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password reset; all sessions revoked.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      description: |
        Requires the current password (wrong current password → `401 InvalidCredentials`). Sets the new
        argon2id hash and revokes every refresh-token family **except** the family of the `vcare_rt` cookie
        sent with this request (the cookie path `/api/auth` covers this route); with no cookie, all families
        are revoked. Any non-suspended account may call it, including `pending` and `rejected` (ADR 0004).
        Rate limit: 5 per 15 min per user.
      x-roles: [patient, doctor, admin]
      x-ownership: self
      x-error-codes: [ValidationFailed, Unauthorized, TokenExpired, InvalidCredentials, AccountSuspended, Conflict, IdempotencyConflict, RateLimited, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed; other sessions revoked.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/UnauthorizedOrInvalidCredentials' }
        '403': { $ref: '#/components/responses/AccountStateForbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get own account
      description: Any non-suspended account, including `pending` and `rejected` (ADR 0004).
      x-roles: [patient, doctor, admin]
      x-ownership: self
      x-error-codes: [Unauthorized, TokenExpired, AccountSuspended, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: The caller's account.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccountStateForbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update own profile fields
      description: |
        Only `fullName`, `phone`, `avatarUrl`, `timezone`, `locale` may change. Any other property —
        including `email`, `role`, `status` — is rejected with `400 ValidationFailed` (unknown properties are
        forbidden). `timezone` must be a valid IANA zone name or `UTC` (fixed offsets such as `+01:00` are rejected); `locale` a BCP-47 tag. Any non-suspended account,
        including `pending` and `rejected` (ADR 0004).
      x-roles: [patient, doctor, admin]
      x-ownership: self
      x-error-codes: [ValidationFailed, Unauthorized, TokenExpired, AccountSuspended, InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: Updated account.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControlNoStore' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/AccountStateForbidden' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── users (admin) ────────────────────────────────────────────────────────────
...
  /.well-known/jwks.json:
    get:
      operationId: getJwks
      tags: [keys]
      summary: Public signing keys (JWKS)
      description: |
        Current and previous Ed25519 public keys (RFC 7517 / RFC 8037). Returned as a bare JWK Set, **not**
        wrapped in the success envelope, so standard JOSE libraries can consume it. Cached
        (`Cache-Control: public, max-age=300`). Budget p95 < 10 ms.
      security: []
      x-roles: [public]
      x-ownership: none
      x-error-codes: [InternalError]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: JWK Set.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            Cache-Control:
              description: Consumers may cache the key set for 5 minutes.
              schema: { type: string, const: 'public, max-age=300' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Jwks' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerUser:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        User access token. EdDSA (Ed25519), `kid` header, TTL 900 s. Claims: `iss=vcare-identity`,
        `aud=["vcare-identity","vcare-care"]`, `sub=<user id as string>`, `typ=user`, `role`, `status`,
        `ev` (email verified; always `true` for accounts created by `register/complete`, ADR 0006), `iat`,
        `exp`, `jti`. `status` may be `pending` or `rejected` (ADR 0004). Never in a cookie or URL.
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Service token from `POST /internal/auth/token`. EdDSA, TTL 300 s. Claims: `iss=vcare-identity`,
        `sub=<client_id>`, `typ=service`, `aud` (must include `vcare-identity` for these routes),
        `scope` (space-separated), `jti`. A `typ=user` token is rejected with `401 ServiceTokenRequired`.
    refreshCookie:
      type: apiKey
      in: cookie
      name: vcare_rt
      description: |
        Opaque 256-bit refresh token (only its sha256 is stored). `HttpOnly; Secure; SameSite=Strict;
        Path=/api/auth; Max-Age=2592000`. Rotated on every use.
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: Caller-supplied request id (UUID). Non-UUID values are replaced by a generated id. Echoed on every response.
      schema: { type: string, format: uuid }
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID. Stored 24 h keyed by (route, principal-or-ip, key) with a body hash.
        Missing → 400 ValidationFailed with details[].field = "Idempotency-Key".
      schema: { type: string, format: uuid }
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional UUID. Stored 24 h keyed by (route, principal-or-ip, key) with a body hash.
      schema: { type: string, format: uuid }
...
  headers:
    XRequestId:
      description: The request id adopted or generated for this request.
      schema: { type: string, format: uuid }
    CacheControlNoStore:
      description: |
        The response must not be stored by any cache: every /api/auth/* response (including malformed-JSON 400s
        and the 404 envelope answered to OPTIONS; in local development a CORS preflight 204), every token
        response, and the health probes.
      schema: { type: string, const: no-store }
...
  responses:
    BadRequest:
      description: Validation failed.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            ValidationFailed:
              value:
                success: false
                error:
                  code: ValidationFailed
                  message: Request validation failed
                  details:
                    - field: email
                      issue: must be an email
                  requestId: 7f1c2a9e-3b4d-4e5f-8a6b-1c2d3e4f5a6b
...
    Conflict:
      description: |
        Conflicts with existing state: a concurrent registration for the same email won the race
        (`register/complete`), or a request with the same `Idempotency-Key` is still in flight (any POST that
        accepts the key). The in-flight case carries `Retry-After: 1` and may be retried after that delay.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            Conflict:
              value:
                success: false
                error: { code: Conflict, message: Email is already registered, details: [], requestId: 7f1c2a9e-3b4d-4e5f-8a6b-1c2d3e4f5a6b }
...
    IdempotencyConflict:
      description: Same Idempotency-Key reused with a different request body.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            IdempotencyConflict:
              value:
                success: false
                error: { code: IdempotencyConflict, message: Idempotency-Key was used with a different request body, details: [], requestId: 7f1c2a9e-3b4d-4e5f-8a6b-1c2d3e4f5a6b }
    RateLimited:
      description: Rate limiter tripped.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            RateLimited:
              value:
                success: false
                error: { code: RateLimited, message: Too many requests, details: [], requestId: 7f1c2a9e-3b4d-4e5f-8a6b-1c2d3e4f5a6b }
    InternalError:
      description: Unhandled error. No stack trace or internals in the body.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            InternalError:
              value:
                success: false
                error: { code: InternalError, message: Internal server error, details: [], requestId: 7f1c2a9e-3b4d-4e5f-8a6b-1c2d3e4f5a6b }
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Base success envelope. Operations narrow `data` (and `meta` for lists) with allOf.
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data:
          description: Operation-specific payload.
        meta:
          type: object
          description: Present only when the operation returns metadata (e.g. pagination).
...
    RegisterStartRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254, examples: [amira.patient@example.test] }
    RegisterCompleteRequest:
      type: object
      additionalProperties: false
      required: [email, code, password, fullName, role, timezone, locale]
      properties:
        email: { type: string, format: email, maxLength: 254, examples: [amira.patient@example.test] }
        code: { $ref: '#/components/schemas/RegistrationCode' }
        password: { $ref: '#/components/schemas/Password' }
        fullName: { type: string, minLength: 1, maxLength: 120, pattern: '\S' }
        role: { type: string, enum: [patient, doctor], description: admin is never accepted. }
        phone: { type: string, pattern: '^\+[1-9][0-9]{7,14}$' }
        timezone: { type: string, minLength: 1, maxLength: 64, description: 'IANA zone name (e.g. Africa/Cairo) or UTC, stored in canonical form; fixed UTC offsets such as +01:00 or GMT+1 are rejected', examples: [Africa/Cairo] }
        locale: { type: string, minLength: 2, maxLength: 35, description: 'BCP-47 tag; stored in canonical form', examples: [ar-EG] }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, minLength: 1, maxLength: 128 }
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
...